+++ src/Runtime/DdevRuntimeHost.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Gaambo\DeployerWordpress\Runtime;
+
+/** @deprecated Use \Gaambo\DeployerUtils\Runtime\DdevRuntimeHost. */
+class DdevRuntimeHost extends \Gaambo\DeployerUtils\Runtime\DdevRuntimeHost
+{
+}